hestia/app/requirements.txt
flask
flask-assets
pyyaml
requests
